requirements_dev.txt
chardet==3.0.4
pdfminer.six==20181108
pytest==6.2.4
twine==3.4.1
